Move helper functions into imports/helpers.py
