+++ ddionrails/instruments/models/instrument.py
@@ -7,9 +7,9 @@ from django.db import models
 from django.urls import reverse
 
 from config.validators import validate_lowercase
-from ddionrails.base.helpers import hash_with_namespace_uuid
 from ddionrails.base.mixins import ModelMixin
 from ddionrails.concepts.models import AnalysisUnit, Period
+from ddionrails.imports.helpers import hash_with_namespace_uuid
 from ddionrails.studies.models import Study
